requirements.txt
rembg==2.0.50
pillow==10.0.0